Add ability to disable CLONE for cluster with async replication
This commit adds the ability to disable CLONE operation during
bootstrap of a MySQL node from a cluster with async replication. This
behaviour is controlled by the CLONE_REQUIRED_ON_ASYNC_CLUSTER boolean
environemnt variable.

The CLONE operation exists to clone the primary database instance to
the secondary before it will join the cluster. The operation itself
might be expensive in a case of big database and moreover dangerous as
it CLONE can't be executed between two MySQL instances with different
versions.