Add ability to set 'wait time' for MySQL cluster upgrade
This commit adds ability to set wait time between MySQL instances
restart during upgrade of the MySQL cluster. The behaviour is
controlled by the integer (in minutes)
REPLICATION_SYNC_WAIT_TIME_ON_UPGRADE environemnt varaible.

The main reason to wait some time between MySQL instances restart
during upgrade is that after the secondary instance is restarted it
needs some time to sync with the primary. If they are both in sync
after the restart of the primary and secondary instances it may lead
to the situation when the orchestrator will fail to promote secondary
to new primary because of replication lag.